requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
requests==2.32.4
python-decouple==3.8
